Dedupe vulns across images too
If the same vulnerability is found in different images,
aggregate them together. This previously relied on the
fingerprint being identical but that intentionally
varies between artifact.

This has the effect of not being able to tie affected
packages back to which particular image uses them (you
just get a list of both), but I think it's worth it
for a much cleaner list.